@@ -1,3 +1,3 @@
 [tool.black]
-target-version = ['py36']
+target-version = ['py37']
 
