@@ -94,7 +98,7 @@ setup(
     ),
     package_dir={"": "src"},
     include_package_data=True,
-    python_requires=">=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*,!=3.5.*",
+    python_requires=">=3.7",
     install_requires=requires,
     tests_require=tests_require,
     extras_require={
