@@ -4,8 +4,8 @@ from typing import List
 from typing import Tuple
 
 # pypi
-from sqlalchemy import Column
 from sqlalchemy.ext.compiler import compiles
+from sqlalchemy.orm import Mapped
 from sqlalchemy.sql import expression
 import sqlalchemy.types
 
@@ -401,6 +401,7 @@ class Acme_Status_Authorization(_Acme_Status_All):
 
     OPTIONS_DEACTIVATE = (
         "pending",
+        "valid",  # a valid auth can be deactivated to uncache it
         "*discovered*",
     )
     OPTIONS_POSSIBLY_PENDING = (
@@ -1006,7 +1007,7 @@ class PrivateKeyType(_mixin_mapping):
 
 
 class _mixin_OperationsEventType(object):
-    operations_event_type_id: Column
+    operations_event_type_id: Mapped[int]
 
     @property
     def event_type_text(self) -> str:
