@@ -42,7 +42,7 @@ class FunctionalTests_Main(unittest.TestCase):
     """
 
     _settings = None
-    _app_settings = None
+    _app_settings: ApplicationSettings
 
     def setUp(self):
         self._settings = settings = get_appsettings(TEST_INI, name="main")
@@ -73,6 +73,7 @@ class FunctionalTests_Main(unittest.TestCase):
     @unittest.skipUnless(RUN_NGINX_TESTS, "Not Running Against: nginx")
     def test_nginx_status(self):
         nginx_servers = self._app_settings.get("nginx.servers_pool")
+        assert isinstance(nginx_servers, list)  # this BETTER be populated!
         nginx_userpass = self._app_settings.get("nginx.userpass", "")
         auth = None
         if nginx_userpass:
@@ -90,6 +91,7 @@ class FunctionalTests_Main(unittest.TestCase):
     @unittest.skipUnless(RUN_NGINX_TESTS, "Not Running Against: nginx")
     def test_nginx_expire_all(self):
         nginx_servers = self._app_settings.get("nginx.servers_pool")
+        assert isinstance(nginx_servers, list)  # this BETTER be populated!
         nginx_userpass = self._app_settings.get("nginx.userpass", "")
         auth = None
         if nginx_userpass:
@@ -108,6 +110,7 @@ class FunctionalTests_Main(unittest.TestCase):
     @unittest.skipUnless(RUN_NGINX_TESTS, "Not Running Against: nginx")
     def test_nginx_expire_domain(self):
         nginx_servers = self._app_settings.get("nginx.servers_pool")
+        assert isinstance(nginx_servers, list)  # this BETTER be populated!
         nginx_userpass = self._app_settings.get("nginx.userpass", "")
         auth = None
         if nginx_userpass:
